+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/Attention.scala
@@ -13,14 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.rnn
-
+package com.intel.analytics.bigdl.nn
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.{Sequential, keras, _}
-import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity, TensorModule}
+import com.intel.analytics.bigdl.nn.abstractnn.TensorModule
+import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
-import com.intel.analytics.bigdl.utils.{Shape, T}
 
 import scala.reflect.ClassTag
 
@@ -35,8 +33,8 @@ private[nn] class AttentionLayer[T: ClassTag](
  (implicit ev: TensorNumeric[T]) extends BaseModule[T] {
 
   override def buildModel(): Module[T] = {
-    // InputX with shape [batch_size, length_x, hidden_size].
-    // InputY with shape [batch_size, length_x, hidden_size]
+    // InputX with shape (batch_size, length_x, hidden_size).
+    // InputY with shape (batch_size, length_x, hidden_size)
     // for self attention, InputX and InputY should be the same.
     // Bias is attention bias that will be added to the result of the dot product.
     val inputX = Input()
@@ -67,7 +65,7 @@ private[nn] class AttentionLayer[T: ClassTag](
       Dropout(initP = (1.0 - attentionDropout)).inputs(softMax)
     } else softMax
     val matmulNoTrans = MM().inputs(drop, contiguousV)
-    // Recombine heads --> [batch_size, length, hidden_size]
+    // Recombine heads --> (batch_size, length, hidden_size)
     val combineHeads = new CombineHeads().inputs(matmulNoTrans)
     // Run the combined outputs through another linear projection layer.
     val outputLayer = TransformerOperation.dense(
@@ -75,6 +73,80 @@ private[nn] class AttentionLayer[T: ClassTag](
     val graph = Graph(Array(inputX, inputY, inputBias), Array(outputLayer))
     if (this.train) graph.training() else graph.evaluate()
     graph
+  }
+}
+// Combine tensor that has been splitted.
+//  input should be tensor with shape (batch_size, num_heads, length, hidden_size/num_heads)
+// output should be tensor with shape (batch_size, length, hidden_size)
+private[nn] class CombineHeads[T: ClassTag](implicit ev: TensorNumeric[T])
+  extends TensorModule[T] {
+
+  private val permutations: (Int, Int) = (2, 3)
+
+  override def updateOutput(input: Tensor[T]): Tensor[T] = {
+    val batchSize = input.size(1)
+    val length = input.size(3)
+    val hiddenSize = input.size(2) * input.size(4)
+
+    output.resizeAs(input).copy(input)
+    output = output.transpose(permutations._1, permutations._2)
+      .reshape(Array(batchSize, length, hiddenSize))
+    output
+  }
+
+  override def updateGradInput(input: Tensor[T], gradOutput: Tensor[T]): Tensor[T] = {
+    val size = Array(input.size(1), input.size(3), input.size(2), input.size(4))
+    if (gradOutput.isContiguous()) {
+      gradInput = gradOutput.view(size)
+    } else {
+      gradInput = gradOutput.contiguous().view(size)
+    }
+    gradInput = gradInput.transpose(permutations._1, permutations._2).contiguous()
+    gradInput
+  }
+}
+
+/**
+ * Split x into different heads, and transpose the resulting value.
+ * The tensor is transposed to insure the inner dimensions hold the correct
+ * values during the matrix multiplication.
+ * input with shape (batch_size, length, hidden_size)
+ * output with shape (batch_size, num_heads, length, hidden_size/num_heads)
+ * @param hiddenSize
+ * @param numHeads
+ * @param mul
+ * @tparam T The numeric type in this module parameters
+ */
+private[nn] class SplitHeads[T: ClassTag](hiddenSize: Int, numHeads: Int,
+                                          mul: Boolean = false)(implicit ev: TensorNumeric[T])
+  extends TensorModule[T] {
+
+  private val depth = hiddenSize / numHeads
+  private val value = ev.fromType(math.pow(depth, -0.5))
+  private val permutations: (Int, Int) = (2, 3)
+
+  override def updateOutput(input: Tensor[T]): Tensor[T] = {
+    val batchSize = input.size(1)
+    val length = input.size(2)
+
+    output.resizeAs(input).copy(input)
+    output = output.reshape(Array(batchSize, length, numHeads, depth))
+      .transpose(permutations._1, permutations._2)
+    if (mul) {
+      output.mul(value)
+    }
+    output
+  }
+
+  override def updateGradInput(input: Tensor[T], gradOutput: Tensor[T]): Tensor[T] = {
+    if (mul) {
+      gradInput.resizeAs(gradOutput).zero().add(value, gradOutput)
+    } else {
+      gradInput.resizeAs(gradOutput).copy(gradOutput)
+    }
+    gradInput = gradInput.transpose(permutations._1, permutations._2).contiguous()
+    gradInput.resize(input.size())
+    gradInput
   }
 }
