+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/FeedForwardNetworkSpec.scala
@@ -13,12 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.rnn
+package com.intel.analytics.bigdl.nn
 
-import com.intel.analytics.bigdl.nn.Linear
-import com.intel.analytics.bigdl.nn.keras.Dense
 import com.intel.analytics.bigdl.tensor.Tensor
-import com.intel.analytics.bigdl.utils.{Shape, T, Table}
+import com.intel.analytics.bigdl.utils.{T, Table}
 import org.scalatest.{FlatSpec, Matchers}
 
 class FeedForwardNetworkSpec extends FlatSpec with Matchers {
